requirements.txt
discord.py
flask
yt-dlp #yt-dlp o wavelink
ytmusicapi
python-dotenv